// Frontend/sakai-ng-master/src/app/demo/models/SucursalViewModel.ts
export class Sucursal{
    sucur_Id?:number;
    sucur_Descripcion?:String;
    sucur_Direccion?:String;
    sucur_Telefono?:String;
    munic_Id?: string;
    munic_Descripcion?: string;
    usuarioCreacion?: string;
    usuarioModificacion?: string;
    sucur_UsuarioCreacion?: number;
    sucur_UsuarioModificacion?: number;
    sucur_FechaCreacion?: number;
    sucur_FechaModificacion?: string;
}