// Frontend/sakai-ng-master/src/app/demo/models/MunicipioViewModel.ts
export class Municipio{
    munic_Id?:string;
    munic_Descripcion?:String;
    depar_Id?: string;
    depar_Descripcion?: string;
    munic_UsuarioCreacion?: number;
    munic_UsuarioModificacion?: number;
    munic_FechaCreacion?: number;
    munic_FechaModificacion?: string;
    usuarioCreacion?: string
    usuarioModificacion?: string
}